--- a/include/simfil/value.h
+++ b/include/simfil/value.h
@@ -170,6 +170,11 @@ struct ValueType4CType<ModelNode> {
     static constexpr ValueType Type = ValueType::Object;
 };
 
+template <>
+struct ValueType4CType<ModelNode::Ptr> {
+    static constexpr ValueType Type = ValueType::Object;
+};
+
 template <ValueType>
 struct ValueTypeInfo;
 
@@ -237,6 +242,30 @@ struct ValueAs<ValueType::String>
     }
 };
 
+template <>
+struct ValueAs<ValueType::Object>
+{
+    template <class VariantType>
+    static auto get(const VariantType& v) -> ModelNode::Ptr
+    {
+        if (auto nodePtr = std::get_if<ModelNode::Ptr>(&v))
+            return *nodePtr;
+        return {};
+    }
+};
+
+template <>
+struct ValueAs<ValueType::Array>
+{
+    template <class VariantType>
+    static auto get(const VariantType& v) -> ModelNode::Ptr
+    {
+        if (auto nodePtr = std::get_if<ModelNode::Ptr>(&v))
+            return *nodePtr;
+        return {};
+    }
+};
+
 class Value
 {
 public:
@@ -278,20 +307,33 @@ public:
 
     static auto field(const ModelNode& node) -> Value
     {
-        return {node.type(), node.value(), model_ptr<ModelNode>(node)};
+        const auto type = node.type();
+        if (type == ValueType::Object || type == ValueType::Array) {
+            return Value{type, model_ptr<ModelNode>(node)};
+        } else {
+            return Value{node.value()};
+        }
     }
 
     static auto field(ModelNode&& node) -> Value
     {
-        auto type = node.type();
-        auto value = node.value();
-        return {type, std::move(value), model_ptr<ModelNode>(std::move(node))};
+        const auto type = node.type();
+        if (type == ValueType::Object || type == ValueType::Array) {
+            return {type, model_ptr<ModelNode>(std::move(node))};
+        } else {
+            return Value{node.value()};
+        }
     }
 
     template <class ModelNodeT>
     static auto field(const model_ptr<ModelNodeT>& node) -> Value
     {
-        return {node->type(), node->value(), node};
+        const auto type = node->type();
+        if (type == ValueType::Object || type == ValueType::Array) {
+            return {type, model_ptr<ModelNode>(node)};
+        } else {
+            return Value{node->value()};
+        }
     }
 
     Value(ValueType type)  // NOLINT
@@ -304,17 +346,12 @@ public:
           , value(std::forward<ArgType>(value))
     {}
 
-    Value(ValueType type, ScalarValueType&& value_, ModelNode::Ptr node)
-        : type(type), node(std::move(node))
-    {
-        std::visit([this](auto&& v){value = v;}, value_);
-    }
 
     Value(ScalarValueType&& value_)  // NOLINT
     {
         std::visit([this](auto&& v){
             type = ValueType4CType<std::decay_t<decltype(v)>>::Type;
-            value = v;
+            value = std::forward<decltype(v)>(v);
         }, value_);
     }
 
@@ -363,8 +400,8 @@ public:
             return fn(this->template as<ValueType::TransientObject>());
         case ValueType::Object:
         case ValueType::Array:
-            if (node) [[likely]]
-                return fn(*node);
+            if (auto nodePtr = std::get_if<ModelNode::Ptr>(&value); nodePtr && *nodePtr) [[likely]]
+                return fn(**nodePtr);
             else
                 return fn(NullType{});
         }
@@ -400,10 +437,21 @@ public:
         return scalarVisitor.result;
     }
 
-    /// Get the string_view of this Value if it has one.
     std::string_view const* stringViewValue() noexcept
     {
         return std::get_if<std::string_view>(&value);
+    }
+
+    ModelNode::Ptr const* nodePtr() const noexcept
+    {
+        return std::get_if<ModelNode::Ptr>(&value);
+    }
+
+    ModelNode const* node() const noexcept
+    {
+        if (auto ptr = nodePtr(); ptr)
+            return &**ptr;
+        return nullptr;
     }
 
     ValueType type;
@@ -414,9 +462,8 @@ public:
         double,
         std::string,
         std::string_view,
-        TransientObject> value;
-
-    ModelNode::Ptr node;
+        TransientObject,
+        ModelNode::Ptr> value;
 };
 
 template <typename ResultT, typename ValueT>