The `simfil::Value` type is huge because of the `ModeNode::Ptr` field.
Moving the node field into the variant saves some bytes and improves performance a lot.

This is just a test, as it removes the option to store a scalar + a node in a Value object.

Main:
```
benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Query typeof Recursive                         100             1     1.73898 s 
                                        17.6426 ms    17.5999 ms     17.695 ms 
                                        241.428 us    204.215 us    290.245 us 
                                                                               
Query field id Recursive                       100             1     1.36372 s 
                                        13.9724 ms    13.8896 ms    14.0689 ms 
                                        456.841 us    402.926 us    551.696 us 
                                                                               
Query field id                                 100             1    28.5413 ms 
                                        286.066 us    284.083 us     289.23 us 
                                        12.5867 us     8.7692 us    17.7284 us
```

This PR:
```
benchmark name                       samples       iterations    est run time
                                     mean          low mean      high mean
                                     std dev       low std dev   high std dev
-------------------------------------------------------------------------------
Query typeof Recursive                         100             1     1.69738 s 
                                        16.4374 ms    16.3952 ms    16.4813 ms 
                                        220.149 us    194.271 us    251.362 us 
                                                                               
Query field id Recursive                       100             1     1.11677 s 
                                         11.213 ms    11.1973 ms    11.2294 ms 
                                        81.8348 us    70.2954 us    97.3137 us 
                                                                               
Query field id                                 100             1    25.6478 ms 
                                        253.815 us    252.875 us    255.383 us 
                                        6.06835 us    4.06118 us    9.42665 us
```